requirements.txt
python-binance
python-dotenv
print-color